backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
bson==0.5.10
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pymongo==3.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.12.2
uritemplate==4.1.1
